+++ libraries/TCA9548/TCA9548.h
@@ -2,9 +2,9 @@
 //
 //    FILE: TCA9548.h
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.2.0
+// VERSION: 0.2.1
 //    DATE: 2021-03-16
-// PURPOSE: Library for TCA9548 I2C multiplexer
+// PURPOSE: Arduino Library for TCA9548 I2C multiplexer and compatibles.
 //     URL: https://github.com/RobTillaart/TCA9548
 
 
@@ -12,7 +12,7 @@
 #include "Wire.h"
 
 
-#define TCA9548_LIB_VERSION             (F("0.2.0"))
+#define TCA9548_LIB_VERSION             (F("0.2.1"))
 
 
 //  ERROR CODES (to be elaborated)
@@ -32,15 +32,20 @@ public:
   bool    isConnected();                      //  find multiplexer on I2C bus
   bool    isConnected(uint8_t address);       //  find any address on I2C bus
 
-  //  channel = 0.. 7
+
+  //  channel = 0..channelCount()-1
+  uint8_t channelCount();
   bool    enableChannel(uint8_t channel);     //  enable this channel non exclusive
   bool    disableChannel(uint8_t channel);
   bool    selectChannel(uint8_t channel);     //  enable only this channel
   bool    isEnabled(uint8_t channel);
   bool    disableAllChannels();
 
+
   //  mask = 0x00 .. 0xFF - every bit is a channel.
-  //  note these are set simultaneously.
+  //         although not for derived types.
+  //  note 1:  these are set simultaneously.
+  //  note 2:  these may include interrupt bits (in derived classes)
   bool    setChannelMask(uint8_t mask);
   uint8_t getChannelMask();
 
@@ -62,7 +67,7 @@ protected:
   uint8_t   _resetPin;         //  default not set == -1 (255)
   bool      _forced;
   int       _error;
-  //  uint8_t   _channels;     //  future PCA954x support.
+  uint8_t   _channels;         //  PCA954x support.
 };
 
 
@@ -70,10 +75,45 @@ protected:
 //
 //  DERIVED CLASS
 //
+//  see - https://github.com/RobTillaart/TCA9548/issues/15
+//
+
+//  SWITCH, 8 channel + reset
 class PCA9548 : public TCA9548
 {
 public:
   PCA9548(uint8_t deviceAddress, TwoWire *wire = &Wire);
+};
+
+
+//  SWITCH, 4 channel + reset
+class PCA9546 : public TCA9548
+{
+public:
+  PCA9546(uint8_t deviceAddress, TwoWire *wire = &Wire);
+  uint8_t getChannelMask();
+};
+
+
+//  DEVICES WITH INTERRUPT 
+
+//  SWITCH, 4 channel + reset + interrupt
+class PCA9545 : public TCA9548
+{
+public:
+  PCA9545(uint8_t deviceAddress, TwoWire *wire = &Wire);
+  uint8_t getChannelMask();
+  uint8_t getInterruptMask();
+};
+
+
+//  SWITCH, 2 channel + reset + interrupt
+class PCA9543 : public TCA9548
+{
+public:
+  PCA9543(uint8_t deviceAddress, TwoWire *wire = &Wire);
+  uint8_t getChannelMask();
+  uint8_t getInterruptMask();
 };